+++ crates/ingestor/src/lib.rs
@@ -1,5 +1,3 @@
-pub mod debug;
-pub mod epoch;
 pub mod error;
 pub mod fetcher;
 pub mod filters;
